+++ permission_handler_android/android/src/main/java/com/baseflow/permissionhandler/PermissionManager.java
@@ -74,6 +74,15 @@ final class PermissionManager implements PluginRegistry.ActivityResultListener, 
             return false;
         }
 
+        // The [onActivityResult] with a [requestResult] that is `null` when the Application was
+        // terminated while not in the foreground with a permission request in progress (e.g. when
+        // Android decides to kill apps while requesting one of the special permissions). In these
+        // cases we should casually return and make sure the `pendingRequestCount` is set to `0`.
+        if (requestResults == null) {
+            pendingRequestCount = 0;
+            return false;
+        }
+
         int status, permission;
 
         if (requestCode == PermissionConstants.PERMISSION_CODE_IGNORE_BATTERY_OPTIMIZATIONS) {
@@ -251,6 +260,12 @@ final class PermissionManager implements PluginRegistry.ActivityResultListener, 
                 }
 
                 requestResults.put(permission, permissionStatus);
+            // [grantResults] can only contain PermissionConstants.PERMISSION_STATUS_GRANTED or PermissionConstants.PERMISSION_STATUS_DENIED status.
+            // But these permissions can have status PermissionConstants.PERMISSION_STATUS_LIMITED, so we need to recheck status
+            } else if (permission == PermissionConstants.PERMISSION_GROUP_PHOTOS || permission == PermissionConstants.PERMISSION_GROUP_VIDEOS) {
+                requestResults.put(
+                    permission,
+                    determinePermissionStatus(permission));
             } else if (!requestResults.containsKey(permission)) {
                 requestResults.put(
                     permission,
@@ -537,7 +552,20 @@ final class PermissionManager implements PluginRegistry.ActivityResultListener, 
                     } else {
                         permissionStatuses.add(PermissionConstants.PERMISSION_STATUS_GRANTED);
                     }
-                } else {
+                } else if (permission ==  PermissionConstants.PERMISSION_GROUP_PHOTOS || permission ==  PermissionConstants.PERMISSION_GROUP_VIDEOS){
+                    final int permissionStatus = ContextCompat.checkSelfPermission(context, name);
+                    int permissionStatusLimited = permissionStatus;
+                    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.UPSIDE_DOWN_CAKE) {
+                        permissionStatusLimited = ContextCompat.checkSelfPermission(context, Manifest.permission.READ_MEDIA_VISUAL_USER_SELECTED);
+                    }
+                    if (permissionStatusLimited == PackageManager.PERMISSION_GRANTED && permissionStatus == PackageManager.PERMISSION_DENIED) {
+                        permissionStatuses.add(PermissionConstants.PERMISSION_STATUS_LIMITED);
+                    } else if (permissionStatus == PackageManager.PERMISSION_GRANTED) {
+                        permissionStatuses.add(PermissionConstants.PERMISSION_STATUS_GRANTED);
+                    }else {
+                        permissionStatuses.add(PermissionUtils.determineDeniedVariant(activity, name));
+                    }
+                }else {
                     final int permissionStatus = ContextCompat.checkSelfPermission(context, name);
                     if (permissionStatus != PackageManager.PERMISSION_GRANTED) {
                         permissionStatuses.add(PermissionUtils.determineDeniedVariant(activity, name));