+++ permission_handler_android/android/src/main/java/com/baseflow/permissionhandler/PermissionUtils.java
@@ -1,6 +1,7 @@
 package com.baseflow.permissionhandler;
 
 import android.Manifest;
+import android.annotation.TargetApi;
 import android.app.Activity;
 import android.content.Context;
 import android.content.SharedPreferences;
@@ -99,6 +100,7 @@ public class PermissionUtils {
         }
     }
 
+    @TargetApi(22)
     static List<String> getManifestNames(Context context, @PermissionConstants.PermissionGroup int permission) {
         final ArrayList<String> permissionNames = new ArrayList<>();