+++ permission_handler_html/lib/permission_handler_html.dart
@@ -1,5 +1,7 @@
-import 'dart:html' as html;
 import 'dart:async';
+import 'dart:js_interop_unsafe';
+
+import 'package:web/web.dart' as web;
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter_web_plugins/flutter_web_plugins.dart';
@@ -9,11 +11,22 @@ import 'web_delegate.dart';
 
 /// Platform implementation of the permission_handler Flutter plugin.
 class WebPermissionHandler extends PermissionHandlerPlatform {
-  static final html.MediaDevices? _devices = html.window.navigator.mediaDevices;
-  static final html.Geolocation _geolocation =
-      html.window.navigator.geolocation;
-  static final html.Permissions? _htmlPermissions =
-      html.window.navigator.permissions;
+  static final web.MediaDevices? _devices = (() {
+    if (!web.window.navigator.has('mediaDevices')) {
+      return null;
+    }
+    return web.window.navigator.mediaDevices;
+  })();
+  static final web.Geolocation _geolocation = web.window.navigator.geolocation;
+  static final web.Permissions? _htmlPermissions = (() {
+    // Using unsafe interop to check availability of `permissions`.
+    // It's not defined as nullable, so merely loading it into a web.Permission? variable
+    // causes the null-check to fail
+    if (!web.window.navigator.has("permissions")) {
+      return null;
+    }
+    return web.window.navigator.permissions;
+  })();
 
   final WebDelegate _webDelegate;