+++ permission_handler_platform_interface/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 4.2.3
+
+* Fixes class name references in the API documentation.
+
+## 4.2.2
+
+* Adds limited access permission for Android 14+.
+
 ## 4.2.1
 
 * Resolves an error that occurred when calling the `shouldShowRequestPermissionRationale` on iOS.